toggleTableView
